% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_city_code.R
\docType{data}
\name{br_city_code}
\alias{br_city_code}
\title{City code according to National Institute of Geography and Statistics}
\format{A data frame with 5570 rows and 7 variables}
\source{
\url{https://github.com/luizpedone/municipal-brazilian-geodata}
}
\usage{
br_city_code
}
\description{
A dataset containing the state cities' names, their respective codes and states.
}
\details{
\itemize{
  \item city_code. a seven digit city code
  \item city_name. city official name
  \item uf_code. a tow digit federal unity (states and the federal district) code
  \item uf. federal unity acronym
  \item uf_name. federal unity name
  \item latitude. city latitude
  \item longitude. city longitude
  ...
}
}
\keyword{datasets}
