% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_indicators_code.R
\docType{data}
\name{br_indicators_code}
\alias{br_indicators_code}
\title{State code according to National Institute of Geography and Statistics}
\format{A data frame with 28 rows and 2 variables}
\source{
\url{https://concla.ibge.gov.br/classificacoes/por-tema/codigo-de-areas/codigo-de-areas.html}
}
\usage{
br_indicators_code
}
\description{
A dataset containing the indicators' names and their respective codes.
}
\details{
\itemize{
  \item indicator. The indicator's name
  \item code. The indicator's number
  ...
}
}
\keyword{datasets}
