% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_state_code.R
\docType{data}
\name{br_state_code}
\alias{br_state_code}
\title{State code according to National Institute of Geography and Statistics}
\format{A data frame with 27 rows and 2 variables}
\source{
\url{https://concla.ibge.gov.br/classificacoes/por-tema/codigo-de-areas/codigo-de-areas.html}
}
\usage{
br_state_code
}
\description{
A dataset containing the state name acronym and their respective codes.
}
\details{
\itemize{
  \item uf_code. a two digits code
  \item uf. federal unity or uf or state acronym
  ...
}
}
\keyword{datasets}
