% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_code.R
\docType{data}
\name{city_code}
\alias{city_code}
\title{City code according to National Institute of Geography and Statistics}
\format{A data frame with 5560 rows and 2 variables}
\source{
\url{https://concla.ibge.gov.br/classificacoes/por-tema/codigo-de-areas/codigo-de-areas.html}
}
\usage{
city_code
}
\description{
A dataset containing the state cities' names and their respective codes.
}
\details{
\itemize{
  \item code. a two digits code
  \item city. federal unity or uf or state acronym
  ...
}
}
\keyword{datasets}
