% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_indicators.R
\docType{data}
\name{df_indicators}
\alias{df_indicators}
\title{State code according to National Institute of Geography and Statistics}
\format{A data frame with 28 rows and 2 variables}
\source{
\url{https://concla.ibge.gov.br/classificacoes/por-tema/codigo-de-areas/codigo-de-areas.html}
}
\usage{
df_indicators
}
\description{
A dataset containing the indicators' names and their respective ids.
}
\details{
\itemize{
  \item indicator. The indicator's name
  \item id. The indicator's number
  ...
}
}
\keyword{datasets}
