% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_replace.R
\name{fuzzy_replace}
\alias{fuzzy_replace}
\title{Replace a vector strings by strings from another vector using fuzzy logic.}
\usage{
fuzzy_replace(string_vector, string_replacement)
}
\arguments{
\item{string_vector}{The vector with strings to be replaced}

\item{string_replacement}{The vector with usually the correct strings to take}
}
\value{
a vector with the provided string_vector replaced by fuzzy match from
    string_replacement vector.
}
\description{
Replace a vector strings by strings from another vector using fuzzy logic.
}
\details{
This is a helpful function to correct city names according to IBGE
    names, when the list is to long to correct by hand.
}
\examples{
string_vector<-c("Mogi das Cruses","Sao Paulo","CANTA GALLO")
string_replacement<-c("Canta Galo","Mogi das Cruzes","São Paulo")
fuzzy_replace(string_vector,string_replacement)
}
