% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_replace.R
\name{fuzzy_search}
\alias{fuzzy_search}
\title{Search for a word or phrase in a second vector that matches the another word in the first
    vector.}
\usage{
fuzzy_search(x = NULL, y = NULL)
}
\arguments{
\item{x}{The vector with strings to be searched for in the second vector}

\item{y}{The vector with usually the correct strings to be found}
}
\value{
A vector of best matches from y based on the x input vector.
}
\description{
Search for a word or phrase in a second vector that matches the another word in the first
    vector.
}
\details{
This is a helpful function to correct city names according to IBGE
    names, when the list is to long to correct by hand.
}
\examples{
wrong_names<-c("Mogi das Cruses","Sao Paulo","CANTA GALLO")
correct_names<-c("Canta Galo","Mogi das Cruzes","São Paulo")
fuzzy_search(x=wrong_names,y=correct_names)
}
