% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_indicators.R
\name{ibge_cidades}
\alias{ibge_cidades}
\title{Downloads indicators from Brazilian state municipalities}
\usage{
ibge_cidades(uf = NULL, indicators = NULL)
}
\arguments{
\item{uf}{string with the state acronym}

\item{indicators}{Up to six indicators' numbers}
}
\value{
A tibble with the indicators
}
\description{
Downloads indicators from Brazilian state municipalities
}
\examples{

pop_ac<-ibge_cidades(uf="ac",indicators=25207)
}
