% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seade_imp.R
\name{seade_imp}
\alias{seade_imp}
\title{Get data of Sao Paulo state, cities and regions from Seade Foundation}
\usage{
seade_imp(place, variable, year)
}
\arguments{
\item{place}{code, according to place data.}

\item{variable}{variable code acording to variable data.}

\item{year}{vector with years you want to get the data}
}
\value{
a data frame with place code, place name, parent place,variable code,
     variable name, period that the data is available, unity, year, and variable name.
}
\description{
Get data of Sao Paulo state, cities and regions from Seade Foundation
}
\examples{

seade_imp(2016:2017,place=1,variable=55)

}
