% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_indicators.R
\name{sp_indicators}
\alias{sp_indicators}
\title{Get state, cities and regions demographics from Sao Paulo.}
\usage{
sp_indicators(place, variable, year)
}
\arguments{
\item{place}{code, according to place data.}

\item{variable}{variable code acording to variable data.}

\item{year}{vector with years you want to get the data}
}
\value{
a data frame with place code, place name, parent place,variable code,
     variable name, period that the data is available, unity, year, and variable name.
}
\description{
Get state, cities and regions demographics from Sao Paulo.
}
\details{
all codes for the places and indicators (variables) are available at `sp_place`
     dataset.
}
\examples{

sp_indicators(2016:2017,place=1,variable=55)

}
