% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_place.R
\docType{data}
\name{sp_place}
\alias{sp_place}
\title{Place code according to Fundação Seade from São Paulo}
\format{A data frame with 711 rows and 3 variables}
\source{
\url{http://doc.seade.gov.br/index.php/API-SEADE}
}
\usage{
sp_place
}
\description{
A dataset containing the places code, their names and region. Places mean not only cities,
    but also larger regions and also the whole state.
}
\details{
\itemize{
  \item place_code. 1 to 3 digits place code
  \item place. São Paulo' places names
  \item parent_place. larger region name where the place is located.
  ...
}
}
\keyword{datasets}
