% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_variable.R
\docType{data}
\name{sp_variable}
\alias{sp_variable}
\title{Variable code according to Fundação Seade from São Paulo}
\format{A data frame with 98 rows and 2 variables}
\source{
\url{http://doc.seade.gov.br/index.php/API-SEADE}
}
\usage{
sp_variable
}
\description{
A dataset containing the variables' codes and ther respective name (in Portuguese)
}
\details{
\itemize{
  \item var_code. 1 to 2 digits variable code
  \item var_name. Variable names in Portuguese
  ...
}
}
\keyword{datasets}
