% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_city_code.R
\docType{data}
\name{br_city_code}
\alias{br_city_code}
\title{City code according to National Institute of Geography and Statistics}
\format{A data frame with 5560 rows and 4 variables}
\source{
\url{https://concla.ibge.gov.br/classificacoes/por-tema/codigo-de-areas/codigo-de-areas.html}
}
\usage{
br_city_code
}
\description{
A dataset containing the state cities' names, their respective codes and states.
}
\details{
\itemize{
  \item city_code. a six digit city code
  \item city_name. city official name
  \item uf_code. a tow digit federal unity (states and the federal distric) code
  \item uf. federal unity name
  ...
}
}
\keyword{datasets}
